+++ packages/dso-toolkit/src/components/label-group/label-group.models.ts
@@ -2,4 +2,5 @@ import { Label } from "../label/label.models.js";
 
 export interface LabelGroup {
   labels: Label[];
+  slotName?: string;
 }