+++ storybook/src/components/label-group/label-group.css-template.ts
@@ -1,5 +1,6 @@
 import { LabelGroup } from "dso-toolkit";
 import { html } from "lit-html";
+import { ifDefined } from "lit-html/directives/if-defined.js";
 
 import { ComponentImplementation } from "../../templates";
 
@@ -7,8 +8,8 @@ export const cssLabelGroup: ComponentImplementation<LabelGroup> = {
   component: "labelGroup",
   implementation: "html-css",
   template: ({ labelTemplate }) =>
-    function labelGroupTemplate({ labels }) {
-      return html`<div class="dso-label-group">
+    function labelGroupTemplate({ labels, slotName }) {
+      return html`<div class="dso-label-group" slot=${ifDefined(slotName)}>
         <ul>
           ${labels.map((label) => html`<li>${labelTemplate(label)}</li>`)}
         </ul>